requirements.txt
Django==1.3
Markdown==2.0.3
distribute==0.6.15
django-debug-toolbar==0.8.5
django-extensions==0.6
django-messages==0.4.4
django-registration==0.7
django-tagging==0.3.1
django-voting==0.1
psycopg2==2.4.2
wsgiref==0.1.2
yolk==0.4.1
